intermediate folder restructure
